requirements.txt
google-generativeai>=0.3.2
ipywidgets>=8.0.0
IPython>=8.0.0
notebook>=6.5.0
streamlit>=1.35.0

python-dotenv>=1.0.1


